add gunicorn to the Pipfile, change urls
I always use gunicorn and it makes it easier to serve this in production if
it's pulled in automatically.

The new Massassi has different download links.  Instead of calling
download_level.php with a level id, and then getting back an HTML page with a
meta refresh, the new level pages now include a comment inside pointing to the
actual zip file.  This prevents the level download counter from increasing when
the level info page is viewed.